Convert `account.py` into `account` package
This PR splits `account.py` into few more modules, so that we can avoid circular dependencies the more work we do on the account level.
